feat: replace token list
