+++ docs/examples/basic.tsx
@@ -1,7 +1,3 @@
-/**
- * iframe: true // 设置为数值可控制 iframe 高度
- */
-
 import React from 'react';
 import Previewer from 'antd-token-previewer';
