--- docs/panel/index.md
@@ -1,1 +1,0 @@
-<code src="../../src/token-panel/index.tsx" inline />
